feat: add FSP verification to placeOrder, replace maxSize with numChunks
Orders now require a Groth16 file-size proof (FSP) at placement time,
replacing the unverified maxSize with ZK-verified numChunks. This ensures
escrow, capacity, and slashing are all based on a cryptographically
proven chunk count.

Key changes:
- MarketStorage: FileOrder.maxSize → numChunks (uint32), add FspVerifier
- MarketOrders.placeOrder: verify FSP proof, price per chunk instead of byte
- All capacity/reward/slash math updated to use numChunks × BYTES_PER_CHUNK
- NodeStaking: capacity tracking uses uint32 chunks (was uint64 bytes)
- Update proof fixtures for all circuits (poi, fsp, keyleak)
- Add MarketFSP.t.sol with 6 real-proof integration tests
- Update muri-artifacts submodule to 11b20d5

+++ test/market/MarketViews.t.sol
@@ -39,7 +39,7 @@ contract MarketViewsTest is MarketTestBase {
         (
             uint256[] memory ids,
             address[] memory owners,
-            uint64[] memory sizes,
+            uint32[] memory sizes,
             uint16[] memory periods,
             uint8[] memory replicas,
             uint8[] memory filled,
@@ -81,7 +81,7 @@ contract MarketViewsTest is MarketTestBase {
             address owner_,
             string memory uri_,
             uint256 root_,
-            uint64 size_,
+            uint32 size_,
             uint16 periods_,
             uint8 replicas_,
             uint8 filled_
